# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Meeting Backend API
  version: 0.1.0
  description: REST contract for the Meeting backend. WebSocket signaling is documented separately in websocket-signaling.md.
servers:
  - url: http://localhost:5180
...
paths:
...
  /api/client-logs:
    post:
      tags: [Logs]
      summary: Upload frontend logs
      description: Accepts a batch of structured browser logs. Each log must contain level, message, and time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientLogsRequest"
      responses:
        "202":
          description: Logs accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientLogsResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/register/code:
    post:
      tags: [Auth]
      summary: Request a registration verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterCodeRequest"
      responses:
        "200":
          description: Registration code issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthCodeDelivery"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email is already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Code requested too frequently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/register/verify:
    post:
      tags: [Auth]
      summary: Complete registration with a verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerifyRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Email is not registered for this flow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login/code:
    post:
      tags: [Auth]
      summary: Request a login verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginCodeRequest"
      responses:
        "200":
          description: Login code issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthCodeDelivery"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Email is not registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Code requested too frequently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login/verify:
    post:
      tags: [Auth]
      summary: Complete login with a verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerifyRequest"
      responses:
        "200":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Email is not registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get the current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"
        "401":
          description: Session is missing, invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/meetings:
    post:
      tags: [Meetings]
      summary: Create a meeting
      description: Requires an authenticated session. The backend ignores hostUserId and hostNickname from the request body and binds the meeting to the current login session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMeetingRequest"
      responses:
        "201":
          description: Meeting created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateMeetingResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/meetings/{meetingNumber}:
    get:
      tags: [Meetings]
      summary: Get a meeting snapshot
      parameters:
        - $ref: "#/components/parameters/MeetingNumber"
      responses:
        "200":
          description: Meeting snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMeetingResponse"
        "404":
          description: Meeting not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/meetings/{meetingNumber}/minutes:
    get:
      tags: [Meetings]
      summary: Get the temporary minutes snapshot
      parameters:
        - $ref: "#/components/parameters/MeetingNumber"
        - $ref: "#/components/parameters/ParticipantIDQuery"
      responses:
        "200":
          description: Minutes snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingMinutesResponse"
        "400":
          description: Missing participantId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Meeting or participant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/meetings/{meetingNumber}/join:
    post:
      tags: [Meetings]
      summary: Join a meeting
      description: If the request is authenticated, the backend binds the participant to the current session user and ignores the request body userId. A registered account can only keep one active participant identity in the same meeting; a newer join removes the older non-host participant.
      parameters:
        - $ref: "#/components/parameters/MeetingNumber"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinMeetingRequest"
      responses:
        "201":
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinMeetingResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Meeting password is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Meeting not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/meetings/{meetingNumber}/participants/{participantID}/ice-servers:
    post:
      tags: [Meetings]
      summary: Get runtime ICE servers for a participant
      parameters:
        - $ref: "#/components/parameters/MeetingNumber"
        - $ref: "#/components/parameters/ParticipantID"
      responses:
        "200":
          description: Runtime ICE configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IceServersResponse"
        "401":
          description: The current session does not match the authenticated participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: The current session cannot request ICE servers for this participant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Meeting or participant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ClientLogsRequest:
      type: object
      required: [logs]
      properties:
        logs:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: "#/components/schemas/ClientLogEntry"
...
    ClientLogsResponse:
      type: object
      required: [status, accepted]
      properties:
        status:
          type: string
          example: accepted
        accepted:
          type: integer
          minimum: 1
    AuthRegisterCodeRequest:
      type: object
      required: [email, nickname]
      properties:
        email:
          type: string
        nickname:
          type: string
    AuthLoginCodeRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
    AuthVerifyRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
        code:
          type: string
    AuthCodeDelivery:
      type: object
      required: [email, purpose, expiresAt, resendAfter, deliveryMode]
      properties:
        email:
          type: string
        purpose:
          type: string
          enum: [register, login]
        debugCode:
          type: string
          description: Development-only debug code returned by the current stub delivery path.
        expiresAt:
          type: string
          format: date-time
        resendAfter:
          type: string
          format: date-time
        deliveryMode:
          type: string
    RegisterResponse:
      type: object
      required: [status, user]
      properties:
        status:
          type: string
          example: registered
        user:
          $ref: "#/components/schemas/AuthUser"
...